Remove private key from memory
Prior to this change multiple copies of the private key existed in
memory during the decoding phase. This change attempt to clean the
private key out of memory by locking it into memory on platforms that
support that (Linux and Darwin) and then zeroing the memory associated
with it after it's used.

Known issues:
* There is a race between the x509 functions and placing the mlock. I
  don't believe this can be resolved w/o re implementing the x509
  function.